refactor(server): split client websocket boundary
